Add error raise when duplicated timestamps are present.
When are present two files of one type (migration, feature, seed) with the same
timestamp, then only the first is executed. And there is no warning. But when you
try to generate a list of sql scripts for this type all will be listed.

I've added a method for finding duplicates and error raise to 'SqlScript::find'
method. If you will add two files of one type with the same timestamp, then while
executing or while listing files for this type there will be raised an error.